feature: Create job event when job is aborted
When a job is aborted during task creation, update the job
and create a corresponding job event using the new
updateJobAndCreateJobEvent method.Refactor Manager methods
to use receiver functions for job and cursor
operations. Add a test to verify that a job
event is created when a job is aborted.

PR  (#7)