test(integration): Remove build tags from integration tests and run them in parallel
This removes the build tags from the integration tests. It also sets up the required test containers once and runs the tests in parallel. Several parts of the code were refactored, partly due to linter feedback, which now applies since the build tags have been removed.

testResoncileAfterSec was removed because it proved to be very flaky. This is due to the asynchronous setup of the integration tests, which makes it impossible to guarantee a specific number of tasks sent within a given timeframe.

PR: #25