feat: add ExternalID to TaskRequest and TaskResponse
This change introduces the ExternalID field to both TaskRequest and
TaskResponse structs, as well as their corresponding protobuf messages.
ExternalID serves as an identifier for a Job, allowing for better
correlation and tracking of tasks and responses. All relevant code,
tests, and protobuf files have been updated to support this new field.

PR (#78)