fix: cancel AMQP receive operation when client closes
An AMQP client deadlock occurs during cleanup iff the user doesn't cancel the context, causing a goroutine to block on receiving (holding a read lock) while the close method waits for that same lock to terminate the connection.

This change ensures we cancel any waiting receive operations when closing the client.

PR: #94